Separate out some of the labware classes
