+++ progress/2026-04-09.md
@@ -39,6 +39,12 @@ Added first-pass deck labware support for four physical holder fixtures:
   - new `src/deck/labware/tip_rack.py`
   - new `TipRackYamlEntry` schema variant
   - exact coordinate lookup via `deck.resolve("tip_rack_a.A1")`
+- Split concrete holder files for easier navigation:
+  - `src/deck/labware/tip_holder.py`
+  - `src/deck/labware/tip_disposal.py`
+  - `src/deck/labware/well_plate_holder.py`
+  - `src/deck/labware/vial_holder.py`
+  - cleaned `src/deck/labware/holder.py` down to the shared base layer (`HolderLabware` and `LabwareSlot`)
 - Moved geometry onto the shared `Labware` base model:
   - added `BoundingBoxGeometry` to `src/deck/labware/labware.py`
   - all current concrete labware now populate `labware.geometry`
@@ -73,6 +79,7 @@ Added first-pass deck labware support for four physical holder fixtures:
 - `pytest tests/test_holder_labware.py tests/test_panda_deck_yaml.py -q` -> passed
 - `pytest tests/test_labware.py tests/test_holder_labware.py tests/test_tip_rack_labware.py -q` -> passed
 - `pytest tests/test_panda_deck_yaml.py tests/test_labware.py tests/test_deck.py tests/test_deck_loader.py tests/test_holder_labware.py tests/validation/test_bounds_validation.py tests/setup/test_protocol_setup.py` -> passed
+- `pytest tests/test_panda_deck_yaml.py tests/test_labware.py tests/test_deck.py tests/test_deck_loader.py tests/test_holder_labware.py tests/test_tip_rack_labware.py tests/validation/test_bounds_validation.py tests/setup/test_protocol_setup.py` -> passed
 
 ## Open Questions
