+++ src/deck/labware/__init__.py
@@ -2,13 +2,13 @@ from .labware import BoundingBoxGeometry, Coordinate3D, Labware
 from .holder import (
     HolderLabware,
     LabwareSlot,
-    TipDisposal,
-    TipHolder,
-    VialHolder,
-    WellPlateHolder,
 )
+from .tip_disposal import TipDisposal
+from .tip_holder import TipHolder
 from .tip_rack import TipRack
+from .vial_holder import VialHolder
 from .well_plate import WellPlate, generate_wells_from_offsets
+from .well_plate_holder import WellPlateHolder
 from .vial import Vial
 
 __all__ = [
