+++ src/deck/labware/holder.py
@@ -164,62 +164,3 @@ class HolderLabware(Labware):
                     continue
                 positions[f"{child_name}.{position_id}"] = coord
         return positions
-
-
-class TipHolder(HolderLabware):
-    """Bounding-box model for the tip holder fixture."""
-
-    model_name: str = "tip_holder"
-    length_mm: float = 138.0
-    width_mm: float = 66.0
-    height_mm: float = 22.0
-
-
-class TipDisposal(HolderLabware):
-    """Bounding-box model for the used-tip disposal fixture."""
-
-    model_name: str = "tip_disposal"
-    length_mm: float = 198.0
-    width_mm: float = 62.0
-    height_mm: float = 30.0
-
-
-class WellPlateHolder(HolderLabware):
-    """Holder for a single well plate or slide-mounted plate assembly."""
-
-    model_name: str = "SlideHolder_Top"
-    length_mm: float = 100.0
-    width_mm: float = 155.0
-    height_mm: float = 14.8
-    labware_support_height_mm: float = 10.0
-    labware_seat_height_from_bottom_mm: float = 5.0
-
-    def get_plate_slot(self, slot_id: str = "plate") -> Coordinate3D:
-        return self.get_slot(slot_id)
-
-
-class VialHolder(HolderLabware):
-    """Holder for a linear array of vial slots."""
-
-    model_name: str = "9VialHolder20mL_TightFit"
-    length_mm: float = 36.2
-    width_mm: float = 300.2
-    height_mm: float = 35.1
-    labware_support_height_mm: float = 35.1
-    labware_seat_height_from_bottom_mm: float = 18.0
-    slot_count: int = Field(default=9, description="Maximum number of supported vial slots.")
-
-    @field_validator("slot_count")
-    def _validate_slot_count(cls, value: int) -> int:
-        if value <= 0:
-            raise ValueError("slot_count must be positive.")
-        return value
-
-    @model_validator(mode="after")
-    def _validate_slot_capacity(self) -> "VialHolder":
-        if len(self.slots) > self.slot_count:
-            raise ValueError("slots count must be <= slot_count.")
-        return self
-
-    def get_vial_slot(self, slot_id: str) -> Coordinate3D:
-        return self.get_slot(slot_id)
